requirements.txt
fastapi
uvicorn
requests
google-cloud-texttospeech
python-multipart
python-dotenv
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai
